Update detail scraping and metadata handling for stability and completeness
- Move detail scraping to isolated browser context per listing
- Fix bug in warranty coverage parsing (enumerate unpacking)
- Add safe env var loading and optional auth token warnings
- Introduce stopwatch utility for performance tracking
- Improve selector handling and error logging for detail pages
- Add Carfax and window sticker URL extraction with fallback logic
- Centralize cookie loading and shared helpers in utils.py
- Expand metadata with runtime, scrolls, and warning details
- Ensure scraper works cleanly with or without premium access
