+++ .env.example
@@ -1,0 +1,3 @@
+# .env.example
+SB_DB_AUTH_TOKEN_0=your_token0_here
+SB_DB_AUTH_TOKEN_1=your_token1_here
